12.0.000 (2013-03-17) - IMPORTANT: the database schema is changed. - Constant K_OTP_LOGIN was added to shared/config/tce_config.php to enable One-Time-Password. - Constant K_LOGOUT_URL was added to shared/config/tce_config.php to customize redirect link after logout. - All test statistics were merged on a single page with more information: tests > results. - Test statistics are now also available on the public page. - The SVG statistic graph is now exported on PDF. - Ordering mode options were added on test editor. - A password-specific test feature was added on test editor. - Some problems on test editor were fixed. - A selector page and popup for tests was added. - The history of repeated tests is now preserved. - Performances of test generation were improved. - Some authorization/permission problems were fixed. - "font" and "size" tags were added on tcecode to set the font family and size. - A virtual keyboard was added on some textarea fields. - Navigation buttons on the public interface now reports the number of questions. - Questions can be exported and imported using a tab-separated values format (TSV). - All CSV formats were changed to TSV (tab-separated-values). - Autosave was added for free answers (only supported by newer web browsers). - Some files were renamed. - TCPDF Library was updated to version 6 ad moved to another folder.

+++ admin/code/tce_functions_omr.php
@@ -2,7 +2,7 @@
 //============================================================+
 // File name   : tce_functions_omr.php
 // Begin       : 2011-05-17
-// Last Update : 2012-07-31
+// Last Update : 2012-12-26
 //
 // Description : Functions to import test data from scanned
 //               OMR (Optical Mark Recognition) sheets.
@@ -273,9 +273,23 @@ function F_importOMRTestData($user_id, $date, $omr_testdata, $omr_answers) {
 	}
 	// get test data
 	$testdata = F_getTestData($test_id);
-	// 1. delete previous test data
-	$sqld = 'DELETE FROM '.K_TABLE_TEST_USER.' WHERE testuser_test_id='.$test_id.' AND testuser_user_id='.$user_id.'';
-	if (!$rd = F_db_query($sqld, $db)) {
+	// 1. check if test is repeatable
+	$sqls = 'SELECT test_id FROM '.K_TABLE_TESTS.' WHERE test_id='.$test_id.' AND test_repeatable=\'1\' LIMIT 1';
+	if ($rs = F_db_query($sqls, $db)) {
+		if ($ms = F_db_fetch_array($rs)) {
+			// 1a. update previous test data if repeatable
+			$sqld = 'UPDATE '.K_TABLE_TEST_USER.' SET testuser_status=testuser_status+1 WHERE testuser_test_id='.$test_id.' AND testuser_user_id='.$user_id.' AND test_repeatable=\'1\' AND testuser_status>3';
+			if (!$rd = F_db_query($sqld, $db)) {
+				F_display_db_error();
+			}
+		} else {
+			// 1b. delete previous test data if not repeatable
+			$sqld = 'DELETE FROM '.K_TABLE_TEST_USER.' WHERE testuser_test_id='.$test_id.' AND testuser_user_id='.$user_id.'';
+			if (!$rd = F_db_query($sqld, $db)) {
+				F_display_db_error();
+			}
+		}
+	} else {
 		F_display_db_error();
 	}
 	// 2. create new user's test entry
@@ -299,6 +313,7 @@ function F_importOMRTestData($user_id, $date, $omr_testdata, $omr_answers) {
 	} else {
 		// get inserted ID
 		$testuser_id = F_db_insert_id($db, K_TABLE_TEST_USER, 'testuser_id');
+		F_updateTestuserStat($date);
 	}
 	// 3. create test log entries
 	$num_questions = count($omr_testdata) - 1;
